refactor: tag 도메인 query-key를 tag.api.ts로 통합
- tag.query-keys.ts를 삭제하고 tagKeys를 tag.api.ts로 병합
- folder-architecture.md 표준 구조(project.api.ts와 동일 패턴)에 맞춤
