--- features/tag/tag.query-keys.ts
@@ -1,4 +1,0 @@
-export const tagKeys = {
-  all: ['tag'] as const,
-  received: (userLink: string, projectId?: number) => [...tagKeys.all, 'received', userLink, projectId] as const,
-};
